admin/requirements.txt
gtts
textual
rich
questionary
litellm
python-dotenv
openai
